# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rayzan Mart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Brands
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all brands
      responses:
        "200":
          description: Brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"

    post:
      operationId: createBrand
      tags: [brands]
      summary: Create brand (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    put:
      operationId: updateBrand
      tags: [brands]
      summary: Update brand (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"

    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete brand (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Orders
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons (Admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"

    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"

  # Banners
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List hero banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeroBanner"

    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroBanner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroBanner"

    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Admin
...
  /settings:
    get:
      operationId: getSiteSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

    put:
      operationId: updateSiteSettings
      tags: [settings]
      summary: Update site settings (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Category:
      type: object
      properties:
        id:
          type: string
        nameBn:
          type: string
        nameEn:
          type: string
        parentId:
          type: string
          nullable: true
        slug:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        sortOrder:
          type: integer
        isActive:
          type: boolean
        productCount:
          type: integer
      required: [id, nameBn, nameEn]
    CreateCategoryBody:
      type: object
      properties:
        nameBn:
          type: string
        nameEn:
          type: string
        parentId:
          type: string
        slug:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
      required: [nameBn, nameEn]
    Brand:
      type: object
      properties:
        id:
          type: string
        nameBn:
          type: string
        nameEn:
          type: string
        slug:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        productCount:
          type: integer
      required: [id, nameBn, nameEn]
    CreateBrandBody:
      type: object
      properties:
        nameBn:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
        isActive:
          type: boolean
      required: [nameBn, nameEn]
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        minOrderAmount:
          type: number
          nullable: true
        maxUses:
          type: integer
          nullable: true
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, code, type, value]
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
      required: [code, type, value]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        orderAmount:
          type: number
      required: [code, orderAmount]
    HeroBanner:
      type: object
      properties:
        id:
          type: string
        imageUrl:
          type: string
        titleBn:
          type: string
        titleEn:
          type: string
        subtitleBn:
          type: string
          nullable: true
        subtitleEn:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        orderIndex:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, imageUrl, titleBn, titleEn]
    CreateBannerBody:
      type: object
      properties:
        imageUrl:
          type: string
        titleBn:
          type: string
        titleEn:
          type: string
        subtitleBn:
          type: string
        subtitleEn:
          type: string
        link:
          type: string
        orderIndex:
          type: integer
        isActive:
          type: boolean
      required: [imageUrl, titleBn, titleEn]
...