# lib/api-zod/src/generated/types/heroBanner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rayzan Mart API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HeroBanner {
  id: string;
  imageUrl: string;
  titleBn: string;
  titleEn: string;
  subtitleBn?: string | null;
  subtitleEn?: string | null;
  link?: string | null;
  orderIndex?: number;
  isActive?: boolean;
  createdAt?: string;
}